+++ itcook-api/src/main/java/com/itcook/cooking/api/domains/post/dto/recipe/RecipeImageUrlDto.java
@@ -1,6 +1,6 @@
 package com.itcook.cooking.api.domains.post.dto.recipe;
 
-import com.itcook.cooking.domain.infra.s3.ImageUrlDto;
+import com.itcook.cooking.domain.domains.infra.s3.ImageUrlDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
