+++ itcook-api/src/main/java/com/itcook/cooking/api/domains/post/service/PostValidationUseCase.java
@@ -4,9 +4,9 @@ import com.itcook.cooking.api.domains.post.dto.recipe.RecipeProcessDto;
 import com.itcook.cooking.domain.common.annotation.UseCase;
 import com.itcook.cooking.domain.domains.archive.entity.Archive;
 import com.itcook.cooking.domain.domains.like.entity.Liked;
-import com.itcook.cooking.domain.infra.s3.ImageFileExtension;
-import com.itcook.cooking.domain.infra.s3.ImageUrlDto;
-import com.itcook.cooking.domain.infra.s3.S3PresignedUrlService;
+import com.itcook.cooking.domain.domains.infra.s3.ImageFileExtension;
+import com.itcook.cooking.domain.domains.infra.s3.ImageUrlDto;
+import com.itcook.cooking.domain.domains.infra.s3.S3PresignedUrlService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
