+++ itcook-api/src/main/java/com/itcook/cooking/api/domains/post/service/RecipeUseCase.java
@@ -23,7 +23,7 @@ import com.itcook.cooking.domain.domains.archive.entity.Archive;
 import com.itcook.cooking.domain.domains.user.entity.ItCookUser;
 import com.itcook.cooking.domain.domains.archive.service.ArchiveService;
 import com.itcook.cooking.domain.domains.user.service.UserService;
-import com.itcook.cooking.domain.infra.s3.ImageUrlDto;
+import com.itcook.cooking.domain.domains.infra.s3.ImageUrlDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.transaction.annotation.Transactional;
