+++ itcook-api/src/main/java/com/itcook/cooking/api/domains/post/service/SearchUseCase.java
@@ -14,8 +14,8 @@ import com.itcook.cooking.domain.domains.post.adaptor.PostAdaptor;
 import com.itcook.cooking.domain.domains.post.entity.Post;
 import com.itcook.cooking.domain.domains.post.repository.PostQuerydslRepository;
 import com.itcook.cooking.domain.domains.post.repository.dto.SearchPostDto;
-import com.itcook.cooking.domain.infra.redis.RedisService;
-import com.itcook.cooking.domain.infra.redis.dto.RankingWords;
+import com.itcook.cooking.domain.domains.infra.redis.RedisService;
+import com.itcook.cooking.domain.domains.infra.redis.dto.RankingWords;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.BiPredicate;
