+++ itcook-api/src/main/java/com/itcook/cooking/api/global/security/jwt/service/JwtTokenProvider.java
@@ -11,7 +11,7 @@ import com.itcook.cooking.api.global.security.jwt.dto.TokenDto;
 import com.itcook.cooking.domain.common.errorcode.CommonErrorCode;
 import com.itcook.cooking.domain.common.errorcode.UserErrorCode;
 import com.itcook.cooking.domain.common.exception.ApiException;
-import com.itcook.cooking.domain.infra.redis.RedisService;
+import com.itcook.cooking.domain.domains.infra.redis.RedisService;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.JwtException;
