+++ itcook-domain/src/main/java/com/itcook/cooking/domain/domains/infra/email/EmailHtmlConstant.java
@@ -1,4 +1,4 @@
-package com.itcook.cooking.domain.infra.email;
+package com.itcook.cooking.domain.domains.infra.email;
 
 public class EmailHtmlConstant {
