+++ itcook-domain/src/main/java/com/itcook/cooking/domain/domains/infra/fcm/FcmService.java
@@ -1,0 +1,9 @@
+package com.itcook.cooking.domain.domains.infra.fcm;
+
+import com.itcook.cooking.domain.domains.infra.fcm.dto.FcmSend;
+
+public interface FcmService {
+
+    void sendMessageTo(FcmSend fcmSend);
+
+}
