+++ itcook-domain/src/main/java/com/itcook/cooking/domain/domains/infra/fcm/dto/FcmSend.java
@@ -1,4 +1,4 @@
-package com.itcook.cooking.domain.infra.fcm.dto;
+package com.itcook.cooking.domain.domains.infra.fcm.dto;
 
 import lombok.Builder;
