+++ itcook-domain/src/main/java/com/itcook/cooking/domain/domains/infra/redis/dto/RankingWords.java
@@ -1,4 +1,4 @@
-package com.itcook.cooking.domain.infra.redis.dto;
+package com.itcook.cooking.domain.domains.infra.redis.dto;
 
 import lombok.Builder;
