+++ itcook-domain/src/main/java/com/itcook/cooking/domain/domains/infra/s3/ImageFileExtension.java
@@ -1,4 +1,4 @@
-package com.itcook.cooking.domain.infra.s3;
+package com.itcook.cooking.domain.domains.infra.s3;
 
 import lombok.Getter;
