+++ itcook-domain/src/main/java/com/itcook/cooking/domain/domains/infra/s3/S3PresignedUrlService.java
@@ -1,4 +1,4 @@
-package com.itcook.cooking.domain.infra.s3;
+package com.itcook.cooking.domain.domains.infra.s3;
 
 public interface S3PresignedUrlService {
