+++ itcook-domain/src/main/java/com/itcook/cooking/domain/domains/user/entity/ItCookUser.java
@@ -1,6 +1,6 @@
 package com.itcook.cooking.domain.domains.user.entity;
 
-import static com.itcook.cooking.domain.infra.email.EmailTemplate.PASSWORD_EMAIL;
+import static com.itcook.cooking.domain.domains.infra.email.EmailTemplate.PASSWORD_EMAIL;
 
 import com.itcook.cooking.domain.common.BaseTimeEntity;
 import com.itcook.cooking.domain.common.events.Events;
@@ -17,7 +17,7 @@ import com.itcook.cooking.domain.domains.user.enums.UserBadge;
 import com.itcook.cooking.domain.domains.user.enums.UserRole;
 import com.itcook.cooking.domain.domains.user.enums.UserState;
 import com.itcook.cooking.domain.common.events.user.UserLeavedEvent;
-import com.itcook.cooking.domain.infra.s3.ImageUrlDto;
+import com.itcook.cooking.domain.domains.infra.s3.ImageUrlDto;
 import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.CascadeType;
