+++ itcook-domain/src/main/java/com/itcook/cooking/domain/domains/user/entity/UserImageRegisterService.java
@@ -1,9 +1,9 @@
 package com.itcook.cooking.domain.domains.user.entity;
 
 import com.itcook.cooking.domain.common.annotation.DomainService;
-import com.itcook.cooking.domain.infra.s3.ImageFileExtension;
-import com.itcook.cooking.domain.infra.s3.ImageUrlDto;
-import com.itcook.cooking.domain.infra.s3.S3PresignedUrlService;
+import com.itcook.cooking.domain.domains.infra.s3.ImageFileExtension;
+import com.itcook.cooking.domain.domains.infra.s3.ImageUrlDto;
+import com.itcook.cooking.domain.domains.infra.s3.S3PresignedUrlService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.util.StringUtils;
