--- itcook-domain/src/main/java/com/itcook/cooking/domain/infra/fcm/FcmService.java
@@ -1,9 +1,0 @@
-package com.itcook.cooking.domain.infra.fcm;
-
-import com.itcook.cooking.domain.infra.fcm.dto.FcmSend;
-
-public interface FcmService {
-
-    void sendMessageTo(FcmSend fcmSend);
-
-}
