+++ itcook-domain/src/test/java/com/itcook/cooking/domain/domains/IntegrationTestSupport.java
@@ -1,14 +1,9 @@
 package com.itcook.cooking.domain.domains;
 
-import com.itcook.cooking.DomainTestApplication;
 import com.itcook.cooking.domain.domains.config.TestConfig;
-import com.itcook.cooking.domain.infra.fcm.FcmService;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.TestPropertySource;
 
 @SpringBootTest
 //@ContextConfiguration(classes = DomainTestApplication.class)
