+++ itcook-infra/src/main/java/com/itcook/cooking/infra/email/ses/AwsSesService.java
@@ -6,7 +6,7 @@ import com.amazonaws.services.simpleemail.model.Content;
 import com.amazonaws.services.simpleemail.model.Destination;
 import com.amazonaws.services.simpleemail.model.Message;
 import com.amazonaws.services.simpleemail.model.SendEmailRequest;
-import com.itcook.cooking.domain.infra.email.AuthCodeService;
+import com.itcook.cooking.domain.domains.infra.email.AuthCodeService;
 import com.itcook.cooking.domain.common.events.email.EmailSendEvent;
 import java.nio.charset.StandardCharsets;
 import lombok.RequiredArgsConstructor;
