+++ itcook-infra/src/test/java/com/itcook/cooking/infra/email/AuthCodeServiceTest.java
@@ -2,7 +2,7 @@ package com.itcook.cooking.infra.email;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import com.itcook.cooking.domain.infra.email.AuthCodeService;
+import com.itcook.cooking.domain.domains.infra.email.AuthCodeService;
 import com.itcook.cooking.domain.common.events.email.EmailSendEvent;
 import com.itcook.cooking.infra.mock.FakeEmailSendService;
 import org.junit.jupiter.api.DisplayName;
