+++ itcook-infra/src/test/java/com/itcook/cooking/infra/mock/FakeEmailSendService.java
@@ -1,6 +1,6 @@
 package com.itcook.cooking.infra.mock;
 
-import com.itcook.cooking.domain.infra.email.AuthCodeService;
+import com.itcook.cooking.domain.domains.infra.email.AuthCodeService;
 import com.itcook.cooking.domain.common.events.email.EmailSendEvent;
 
 public class FakeEmailSendService implements AuthCodeService {
